requirements.txt
discord.py
python-dotenv
psycopg[binary,pool]>=3.1.18